ajout guard pour le board le post login de l'api retourne un PlayerEntity si connecter sinon refus de connection

+++ UcmEmpire/src/main/java/com/ucm/ucmempire/dal/repository/BoardRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface BoardRepository extends CrudRepository<BoardEntity , Long> {
+public interface BoardRepository extends CrudRepository<BoardEntity , Integer> {
 
     // @Query(value = "Select b From BoardEntity b JOIN PlayerEntity p ON b.id = p.boardEntity.id Where p.id = :player_id")
     Optional<BoardEntity> findBoardEntityById (Integer id_player);
